// amplify/functions/dynamoDB-function/graphql/API.ts
export type SessionAttended = {
  __typename: "SessionAttended",
  createdAt: string,
  date: string,
  earningsThatSession?: number | null,
  id: string,
  sessionAttendedId: string,
  updatedAt: string,
  user?: User | null,
};

export type User = {
  __typename: "User",
  createdAt: string,
  email: string,
  firstName: string,
  id: string,
  lastName: string,
  role?: string | null,
  sessionsAttended?: ModelSessionAttendedConnection | null,
  totalEarnings?: number | null,
  updatedAt: string,
};

export type ModelSessionAttendedConnection = {
  __typename: "ModelSessionAttendedConnection",
  items:  Array<SessionAttended | null >,
  nextToken?: string | null,
};

export type ModelSessionAttendedFilterInput = {
  and?: Array< ModelSessionAttendedFilterInput | null > | null,
  createdAt?: ModelStringInput | null,
  date?: ModelStringInput | null,
  earningsThatSession?: ModelFloatInput | null,
  id?: ModelIDInput | null,
  not?: ModelSessionAttendedFilterInput | null,
  or?: Array< ModelSessionAttendedFilterInput | null > | null,
  sessionAttendedId?: ModelIDInput | null,
  updatedAt?: ModelStringInput | null,
};

export type ModelStringInput = {
  attributeExists?: boolean | null,
  attributeType?: ModelAttributeTypes | null,
  beginsWith?: string | null,
  between?: Array< string | null > | null,
  contains?: string | null,
  eq?: string | null,
  ge?: string | null,
  gt?: string | null,
  le?: string | null,
  lt?: string | null,
  ne?: string | null,
  notContains?: string | null,
  size?: ModelSizeInput | null,
};
// ...
export type ModelFloatInput = {
  attributeExists?: boolean | null,
  attributeType?: ModelAttributeTypes | null,
  between?: Array< number | null > | null,
  eq?: number | null,
  ge?: number | null,
  gt?: number | null,
  le?: number | null,
  lt?: number | null,
  ne?: number | null,
};

export type ModelIDInput = {
  attributeExists?: boolean | null,
  attributeType?: ModelAttributeTypes | null,
  beginsWith?: string | null,
  between?: Array< string | null > | null,
  contains?: string | null,
  eq?: string | null,
  ge?: string | null,
  gt?: string | null,
  le?: string | null,
  lt?: string | null,
  ne?: string | null,
  notContains?: string | null,
  size?: ModelSizeInput | null,
};

export enum ModelSortDirection {
  ASC = "ASC",
  DESC = "DESC",
}


export type ModelUserFilterInput = {
  and?: Array< ModelUserFilterInput | null > | null,
  createdAt?: ModelStringInput | null,
  email?: ModelStringInput | null,
  firstName?: ModelStringInput | null,
  id?: ModelIDInput | null,
  lastName?: ModelStringInput | null,
  not?: ModelUserFilterInput | null,
  or?: Array< ModelUserFilterInput | null > | null,
  role?: ModelStringInput | null,
  totalEarnings?: ModelFloatInput | null,
  updatedAt?: ModelStringInput | null,
};

export type ModelUserConnection = {
  __typename: "ModelUserConnection",
  items:  Array<User | null >,
  nextToken?: string | null,
};
// ...
export type GetSessionAttendedQueryVariables = {
  id: string,
};

export type GetSessionAttendedQuery = {
  getSessionAttended?:  {
    __typename: "SessionAttended",
    createdAt: string,
    date: string,
    earningsThatSession?: number | null,
    id: string,
    sessionAttendedId: string,
    updatedAt: string,
    user?:  {
      __typename: "User",
      createdAt: string,
      email: string,
      firstName: string,
      id: string,
      lastName: string,
      role?: string | null,
      totalEarnings?: number | null,
      updatedAt: string,
    } | null,
  } | null,
};

export type GetUserQueryVariables = {
  id: string,
};

export type GetUserQuery = {
  getUser?:  {
    __typename: "User",
    createdAt: string,
    email: string,
    firstName: string,
    id: string,
    lastName: string,
    role?: string | null,
    sessionsAttended?:  {
      __typename: "ModelSessionAttendedConnection",
      nextToken?: string | null,
    } | null,
    totalEarnings?: number | null,
    updatedAt: string,
  } | null,
};

export type ListSessionAttendedsQueryVariables = {
  filter?: ModelSessionAttendedFilterInput | null,
  id?: string | null,
  limit?: number | null,
  nextToken?: string | null,
  sortDirection?: ModelSortDirection | null,
};

export type ListSessionAttendedsQuery = {
  listSessionAttendeds?:  {
    __typename: "ModelSessionAttendedConnection",
    items:  Array< {
      __typename: "SessionAttended",
      createdAt: string,
      date: string,
      earningsThatSession?: number | null,
      id: string,
      sessionAttendedId: string,
      updatedAt: string,
    } | null >,
    nextToken?: string | null,
  } | null,
};

export type ListUsersQueryVariables = {
  filter?: ModelUserFilterInput | null,
  id?: string | null,
  limit?: number | null,
  nextToken?: string | null,
  sortDirection?: ModelSortDirection | null,
};

export type ListUsersQuery = {
  listUsers?:  {
    __typename: "ModelUserConnection",
    items:  Array< {
      __typename: "User",
      createdAt: string,
      email: string,
      firstName: string,
      id: string,
      lastName: string,
      role?: string | null,
      totalEarnings?: number | null,
      updatedAt: string,
    } | null >,
    nextToken?: string | null,
  } | null,
};